[FIX] Correção de instância do VueSocketIO ao logar na aplicação.
#169 #190 #186 #155
